Enable empty poll removal by deleting one of its options
Previously, the user had to click on the "poll" icon at the bottom of
the compose form in order to remove a poll. This commit changes how poll
options behave so as to enable an empty poll to be removed through the
deletion of either one of its options.

This closes #317.