Draft: Implement nesting on threads
This is an attempt to implement nesting in order to make a reply to a status
appear nested on its parent.

To do:
* Create the vertical lines;
* Decide on a nesting limit. I think two or three nested statuses seems
reasonable but I'm still not sure how to represent the further nested
statuses;
* Once we decide on a nesting limit, change the new class names accordingly.
The current names are only placeholders.

This commit passed on all tests and I haven't found anything broken yet.
Still, if you find any bugs or problems (or have any suggestions) please
let me know.